fixing css and layout and cleanup, trim down
