--- src/app/head.tsx
@@ -1,7 +1,0 @@
-export default async function Head() {
-  return (
-    <>
-      <title>I am Page!</title>
-    </>
-  );
-}
