--- src/app/layout.tsx
@@ -1,14 +1,0 @@
-export default function RootLayout({
-  children,
-}: {
-  children: React.ReactNode;
-}) {
-  return (
-    <html>
-      <head></head>
-      <body>
-        <main>{children}</main>
-      </body>
-    </html>
-  );
-}
