+++ src/components/Seo.tsx
@@ -1,3 +1,5 @@
+'use-client'
+
 import Head from 'next/head';
 import { useRouter } from 'next/router';
 import React from 'react';
