+++ src/pages/_app.tsx
@@ -1,10 +1,12 @@
 import "../styles/globals.css";
-import { Toaster } from "react-hot-toast";
+import "../styles/page.css";
+import "../styles/aboutus.css"
+import "../styles/playnow.css"
+import "../styles/store.css";
 import type { AppProps } from "next/app";
-import { ThemeProvider } from "next-themes";
-import Head from "next/head";
 import React from "react";
-import VolumeContextProvider from "../components/VolumeContext";
+import Seo from "@/components/Seo";
+import { ThemeProvider } from "next-themes";
 
 function MyApp({
   Component,
@@ -12,18 +14,10 @@ function MyApp({
 }: AppProps): React.ReactElement<AppProps> {
   return (
     <>
-      <Head>
-        {/* https://nextjs.org/docs/messages/no-document-viewport-meta */}
-        <link rel="shortcut icon" href="/favicon.ico" />
-          <meta charSet='utf-8' />
-          <meta name='viewport' content='width=device-width, initial-scale=1' />
-      </Head>
-      <Toaster />
-     <ThemeProvider enableSystem={true} attribute="class">
-        <VolumeContextProvider>
-          <Component {...pageProps} />
-        </VolumeContextProvider>
-    </ThemeProvider>
+      <Seo />
+        <ThemeProvider enableSystem={true} attribute="class">
+   <Component {...pageProps} />
+      </ThemeProvider>
     </>
       );
     }
